+++ src/pydrobert/torch/data.py
@@ -39,15 +39,6 @@ import torch.utils.data
 import param
 import pydrobert.torch
 
-try:
-    str
-except NameError:
-    str = str
-
-__author__ = "Sean Robertson"
-__email__ = "sdrobert@cs.toronto.edu"
-__license__ = "Apache 2.0"
-__copyright__ = "Copyright 2019 Sean Robertson"
 __all__ = [
     "context_window_seq_to_batch",
     "ContextWindowDataParams",
@@ -986,7 +977,7 @@ def transcript_to_token(
     unk: Optional[Union[str, int]] = None,
     skip_frame_times: bool = False,
 ) -> torch.Tensor:
-    """Convert a transcript to a token sequence
+    r"""Convert a transcript to a token sequence
 
     This method converts `transcript` of length ``R`` to a long tensor `tok` of shape
     ``(R, 3)``, the latter suitable as a reference or hypothesis token sequence for an
@@ -1000,13 +991,13 @@ def transcript_to_token(
     times, in seconds, of the token, and will be converted to frames for `tok`. If
     `frame_shift_ms` is unspecified, ``start`` and ``end`` are assumed to already be
     frame times. If ``start`` and ``end`` were unspecified, values of ``-1``,
-    representing unknown, will be inserted into ``r[i, 1:]``
+    representing unknown, will be inserted into ``tok[r, 1:]``
 
     Parameters
     ----------
     transcript : sequence
     token2id : dict, optional
-    frame_shift_ms : float, optional
+    frame_shift_ms : float or None, optional
     unk : str or int, optional
         If not :obj:`None`, specifies the out-of-vocabulary token. If `unk`
         exists in `token2id`, the ``token2id[unk]`` will be used as the
@@ -1021,11 +1012,48 @@ def transcript_to_token(
     -------
     tok : torch.Tensor
 
+    Warnings
+    --------
+    The frame index bounds inferred using `frame_shift_ms` should not be used directly
+    in evaluation. See the below note.
+
     Notes
     -----
     If you are dealing with raw audio, each "frame" is just a sample. The appropriate
     value for `frame_shift_ms` is ``1000 / sample_rate_hz`` (since there are
     ``sample_rate_hz / 1000`` samples per millisecond).
+
+    Converting to frame indices from start and end times follows an overly-simplistic
+    equation. Letting :math:`(s_s, e_s)` be the start and end times in seconds,
+    :math:`(s_f, e_f)` be the corresponding start and end frames, :math:`\Delta` be
+    the frame shift in milliseconds, and :math:`I[\cdot]` be the indicator function.
+    Then
+
+    .. math::
+
+        s_f = floor\left(\frac{1000s_s}{\Delta}\right) \\
+        e_f = \max\left(s_s + I[s_s = e_s],
+                        round\left(\frac{1000e_s}{\Delta}\right)\right)
+
+    For a given token index, ``tok[r, 1] = s_f`` and ``tok[r, 2] = e_f``. ``tok[r, 1]``
+    is supposed to be the inclusive start frame of the segment and ``tok[r, 2]`` the
+    exclusive end frame. :math:`(s_f, e_f)` fail to be these on two accounts. First,
+    they do not consider the frame length. First, while frames may be spaced
+    :math:`\Delta` milliseconds apart, they will usually be overlapping. Because of this
+    overlap, the coefficients of frames :math:`s_f - 1` and :math:`e_f` may be in part
+    dependent on the audio samples within the segment. Second, ignoring frame length,
+    :frac:`e_f = ceil(1000e_s/\Delta)` would be more appropriate for an exclusive upper
+    bound. However, :mod:`pydrobert.speech.compute` (and other, mainstream feature
+    computation packages), the total number of frames in the utterance is calculated as
+    :math:`T_f = ceil(1000T_s/\Delta)`, where :math:`T_s` is the length of the utterance
+    in seconds. The above equation ensures :math:`\max(e_f) \leq T_f`, which is a
+    neccessary criterion for a valid :class:`SpectDataSet` (see
+    :func:`validate_spec_data_set`).
+
+    Accounting for both of these assumptions would involve computing the support of
+    each existing frame in seconds and intersecting that with the provided interval in
+    seconds. As such, the derived frame bounds should not be used for an official
+    evaluation. This function should suffice for most training objectives, however.
     """
     if token2id is not None and unk in token2id:
         unk = token2id[unk]
@@ -1039,9 +1067,14 @@ def transcript_to_token(
             if len(token) == 3 and np.isreal(token[1]) and np.isreal(token[2]):
                 token, start, end = token
                 if frame_shift_ms:
-                    start = (1000 * start) / frame_shift_ms
-                    end = (1000 * end) / frame_shift_ms
-                start, end = int(start), int(end)
+                    if start == end:
+                        start = end = (1000 * start) // frame_shift_ms
+                    else:
+                        start = (1000 * start) // frame_shift_ms
+                        end = (1000 * end + 0.5 * frame_shift_ms) // frame_shift_ms
+                        end = max(end, start + 1)
+                else:
+                    start, end = int(start), int(end)
         except TypeError:
             pass
         if token2id is None:
@@ -1077,6 +1110,12 @@ def token_to_transcript(
     Returns
     -------
     token : list
+
+    Warnings
+    --------
+    The time interval inferred using `frame_shift_ms` is unlikely to be perfectly
+    correct. See the note in :func:`transcript_to_token` for more details about the
+    ambiguity in converting between seconds and frames.
     """
     transcript = []
     for tup in tok: